Add files via upload
The jupyter notebook parses relevant information about apartments in Nur-Sultan city advertised in the popular website krisha.kz in Kazakhstan. After that, the relevant information is stored in a dataset which would be analyzed later. In the end, models are trained based on the data in the dataset to predict prices of apartments by using input information about them.