Add files via upload
The jupyter notebook analyzes house prices in Nur-Sultan city. In the beginning, it gets information about advertised apartments in the popular website in Kazakshtan. Later on, a dataset created by the imported information is analyzed and data is vizualized. In the end, models are created based on the data in the dataset which can be used to predict prices of apartments in Nur-Sultan city.

Be careful when running the whole jupyter. The first part which parses information from the website requires much time. So if you do not want to wait for it, it is better to run from the second part which imports information from a csv file.